Moved billing to billing folder
